+++ src/main/scala/apikeysteward/services/PermissionService.scala
@@ -1,14 +1,14 @@
 package apikeysteward.services
 
 import apikeysteward.model.ApiKeyTemplate.ApiKeyTemplateId
-import apikeysteward.model.ResourceServer.ResourceServerId
 import apikeysteward.model.Permission
 import apikeysteward.model.Permission.PermissionId
-import apikeysteward.model.RepositoryErrors.ApiKeyTemplatesPermissionsDbError.ApiKeyTemplatesPermissionsInsertionError
 import apikeysteward.model.RepositoryErrors.GenericError
-import apikeysteward.model.RepositoryErrors.ResourceServerDbError.ResourceServerNotFoundError
 import apikeysteward.model.RepositoryErrors.PermissionDbError.PermissionInsertionError.PermissionAlreadyExistsError
 import apikeysteward.model.RepositoryErrors.PermissionDbError.{PermissionInsertionError, PermissionNotFoundError}
+import apikeysteward.model.RepositoryErrors.ResourceServerDbError.ResourceServerNotFoundError
+import apikeysteward.model.ResourceServer.ResourceServerId
+import apikeysteward.model.Tenant.TenantId
 import apikeysteward.repositories.{ApiKeyTemplateRepository, PermissionRepository, ResourceServerRepository}
 import apikeysteward.routes.model.admin.permission.CreatePermissionRequest
 import apikeysteward.utils.Retry.RetryException
@@ -16,6 +16,8 @@ import apikeysteward.utils.{Logging, Retry}
 import cats.data.EitherT
 import cats.effect.IO
 import cats.implicits.catsSyntaxEitherId
+
+import java.util.UUID
 
 class PermissionService(
     uuidGenerator: UuidGenerator,
@@ -25,12 +27,14 @@ class PermissionService(
 ) extends Logging {
 
   def createPermission(
+      publicTenantId: TenantId,
       resourceServerId: ResourceServerId,
       createPermissionRequest: CreatePermissionRequest
   ): IO[Either[PermissionInsertionError, Permission]] =
-    createPermissionWithRetry(resourceServerId, createPermissionRequest)
+    createPermissionWithRetry(publicTenantId, resourceServerId, createPermissionRequest)
 
   private def createPermissionWithRetry(
+      publicTenantId: TenantId,
       resourceServerId: ResourceServerId,
       createPermissionRequest: CreatePermissionRequest
   ): IO[Either[PermissionInsertionError, Permission]] = {
@@ -48,7 +52,7 @@ class PermissionService(
 
         _ <- logger.info("Inserting Permission into database...")
         newPermission <- permissionRepository
-          .insert(resourceServerId, Permission.from(permissionId, createPermissionRequest))
+          .insert(publicTenantId, resourceServerId, Permission.from(permissionId, createPermissionRequest))
           .flatTap {
             case Right(_) => logger.info(s"Inserted Permission with permissionId: [$permissionId] into database.")
             case Left(e)  => logger.warn(s"Could not insert Permission into database because: ${e.message}")
@@ -62,45 +66,52 @@ class PermissionService(
   }
 
   def deletePermission(
+      publicTenantId: TenantId,
       resourceServerId: ResourceServerId,
       permissionId: PermissionId
   ): IO[Either[PermissionNotFoundError, Permission]] =
-    permissionRepository.delete(resourceServerId, permissionId).flatTap {
+    permissionRepository.delete(publicTenantId, resourceServerId, permissionId).flatTap {
       case Right(_) => logger.info(s"Deleted Permission with permissionId: [$permissionId].")
       case Left(e) =>
         logger.warn(s"Could not delete Permission with permissionId: [$permissionId] because: ${e.message}")
     }
 
-  def getBy(resourceServerId: ResourceServerId, permissionId: PermissionId): IO[Option[Permission]] =
-    permissionRepository.getBy(resourceServerId, permissionId)
+  def getBy(
+      publicTenantId: TenantId,
+      resourceServerId: ResourceServerId,
+      permissionId: PermissionId
+  ): IO[Option[Permission]] =
+    permissionRepository.getBy(publicTenantId, resourceServerId, permissionId)
 
   def getAllForTemplate(
+      publicTenantId: TenantId,
       templateId: ApiKeyTemplateId
   ): IO[Either[GenericError.ApiKeyTemplateDoesNotExistError, List[Permission]]] =
     (for {
       _ <- EitherT(
         apiKeyTemplateRepository
-          .getBy(templateId)
+          .getBy(publicTenantId, templateId)
           .map(_.toRight(GenericError.ApiKeyTemplateDoesNotExistError(templateId)))
       )
 
       result <- EitherT.liftF[IO, GenericError.ApiKeyTemplateDoesNotExistError, List[Permission]](
-        permissionRepository.getAllFor(templateId)
+        permissionRepository.getAllFor(publicTenantId, templateId)
       )
     } yield result).value
 
   def getAllBy(
+      publicTenantId: TenantId,
       resourceServerId: ResourceServerId
   )(nameFragment: Option[String]): IO[Either[ResourceServerNotFoundError, List[Permission]]] =
     (for {
       _ <- EitherT(
         resourceServerRepository
-          .getBy(resourceServerId)
+          .getBy(publicTenantId, resourceServerId)
           .map(_.toRight(ResourceServerNotFoundError(resourceServerId.toString)))
       )
 
       result <- EitherT.liftF[IO, ResourceServerNotFoundError, List[Permission]](
-        permissionRepository.getAllBy(resourceServerId)(nameFragment)
+        permissionRepository.getAllBy(publicTenantId, resourceServerId)(nameFragment)
       )
     } yield result).value
